feat: add EditPopupComponent for editing data
the code changes introduce a new file `EditPopupComponent.jsx` that implements a popup component for editing data. This component is used to display a form with input fields for each field in the `fields` prop. When the "Actualizar" button is clicked, the form data is collected and passed to the `onSubmit` function along with the `currentElementId`. The purpose of this commit is to add the functionality for editing data in the application.