feat: add handleSendToInvoice function for sending product to invoice
This commit adds the `handleSendToInvoice` function to the `actionHandlers.js` file. The function allows sending a product to the invoice creator by making a POST request to the `/api/invoices/sendProductToInvoiceCreator` endpoint. It retrieves the `idProducto` from the provided row and sends it in the request payload. If the request is successful, the product is added to the invoice and the updated list of products is stored in the session. If there is an error, an error message is displayed. This enhancement improves the functionality of the application by enabling users to easily add products to the invoice.