refactor: update CSS selectors for expand and export buttons
The CSS changes in `product-client-invoice-styles.css` update the selectors for the expand and export buttons in the table. Instead of using the `label[for*="details"]` and `label[for*="export"]` selectors, the new selectors are `#expand_button` and `#export_button` respectively. This change improves the clarity and specificity of the CSS styles for these buttons.