feat: fetch updated invoices from API
This commit adds the `fetchUpdatedInvoices` function to the `actionHandlers.js` file. The function makes a GET request to the `/api/invoices/` endpoint to fetch the updated invoices. If the response is successful, the invoices are formatted using the `formatInvoices` function and set using the `setInvoices` callback. If there is an error, it is logged to the console.