feat: add handleDelete function for deleting invoices
The code changes in `actionHandlers.js` add the `handleDelete` function, which is responsible for deleting an invoice. It makes a DELETE request to the `/api/invoices/delete/{id}` endpoint, where `{id}` is the ID of the invoice to be deleted. If the request is successful, the `fetchUpdatedInvoices` function is called to update the list of invoices.