feat: add fetchDetailsInvoiceInfo function for fetching invoice details
This commit adds the `fetchDetailsInvoiceInfo` function to the `tableHelper.js` file. The function makes a GET request to the `/api/invoices/details/{id}` endpoint, where `{id}` is the ID of the invoice to fetch the details for. If the request is successful, the response data is returned. If there is an error, it is logged to the console.