feat: add fetchDetailsInvoiceInfo function for fetching invoice details
The code changes in `exportHandlers.js` add a new function `fetchDetailsInvoiceInfo` that allows fetching the details of an invoice. This function makes a GET request to the `/api/invoices/details/{invoiceId}` endpoint to retrieve the invoice details for the specified invoice ID. The fetched data is returned as a JSON object.