feat: add exportXML function for exporting invoices as XML
This commit adds a new function `exportXML` to `exportHandlers.js` that allows exporting invoices as XML files. The function fetches the invoice details using the `fetchDetailsInvoiceInfo` function and reformats the data to match the desired XML structure. It then converts the reformatted data to XML using the `toXML` function from the `jstoxml` dependency. The resulting XML is downloaded as a file named `factura.xml`. This functionality enables users to export invoices in XML format.