feat: add selectClient endpoint for selecting a client
This commit adds a new endpoint `/selectClient` to the `FacturasController.java` file. The endpoint allows users to select a client by providing their identification. The selected client is retrieved using the `getClientByIdentificationAndProveedor` method from the `clienteService` and returned as a response. This functionality enables users to easily select clients for invoice creation.