feat: add logoutHandler
the code changes in `logoutHandler.js` add the `logout` function, which sends a GET request to the logout API endpoint with the appropriate credentials. If the response is successful, the session storage is cleared and the user is redirected to the login page. If there is an error, the session storage is still cleared and the user is redirected to the login page, with the error message logged to the console.