feat: add PATCH endpoint for changing provider state
the code changes in `AdminController.java` add a new PATCH endpoint `/changeProviderState` for changing the state of a provider. This endpoint takes an `idProvider` parameter and updates the state of the corresponding provider. If the update is successful, the admin dashboard is returned. If there is an error, a 400 status code is returned.