refactor: prevent adding product to cart with a quantity less than or equal to 0
The addToCart method in FacturasController.java has been updated to check if the quantity of the product being added is greater than 0. If the quantity is less than or equal to 0, an error response is returned with a message indicating that the quantity must be greater than 0. This prevents invalid quantities from being added to the cart.