refactor: handle DataIntegrityViolationException when changing email in UsuarioService
The changeEmail method in UsuarioService.java has been updated to handle the DataIntegrityViolationException that may occur when changing the email of a user. If the exception message contains the word "correo", it throws an IllegalArgumentException with the message "Parece que ya existe una cuenta con este correo.". Otherwise, it throws an IllegalArgumentException with the message "Hubo un error al cambiar el correo. Intenta de nuevo.". This ensures that appropriate error messages are displayed when there is an issue changing the email.